Minor tweaks.

+++ code/modules/fluids/fluid_controller.dm
@@ -6,7 +6,7 @@ var/datum/controller/process/fluids/fluid_master
 
 /datum/controller/process/fluids/setup()
 	name = "fluids"
-	schedule_interval = 5
+	schedule_interval = 3
 	fluid_master = src
 	module_controllers["Fluids"] = fluid_master
 
@@ -18,40 +18,40 @@ var/datum/controller/process/fluids/fluid_master
 	// Process water sources.
 	for(var/thing in water_sources)
 		var/turf/T = thing
-		T.flood_neighbors()
+		if(T) T.flood_neighbors()
 		scheck()
 
  	// Process general fluid spread.
 	var/list/spreading_fluids = active_fluids.Copy()
 	for(var/thing in spreading_fluids)
 		var/obj/effect/fluid/F = thing
-		F.spread()
+		if(F) F.spread()
 		scheck()
 
 	// Equalize fluids.
 	for(var/thing in spreading_fluids)
-		if(!(thing in active_fluids))
-			continue
+		if(!(thing in active_fluids)) continue
 		var/obj/effect/fluid/F = thing
-		F.equalize()
+		if(F) F.equalize()
 		scheck()
 	spreading_fluids.Cut()
 
 	// Update icons!
 	for(var/thing in active_fluids)
 		var/obj/effect/fluid/F = thing
-		if(!F.loc || F.loc != F.start_loc)
-			qdel(F)
-		if(F.fluid_amount <= FLUID_EVAPORATION_POINT && prob(10))
-			F.lose_fluid(rand(1,3))
-		if(F.fluid_amount <= FLUID_DELETING)
-			qdel(F)
-		else
-			F.update_position_and_alpha()
+		if(F)
+			if(!F.loc || F.loc != F.start_loc)
+				qdel(F)
+			if(F.fluid_amount <= FLUID_EVAPORATION_POINT && prob(10))
+				F.lose_fluid(rand(1,3))
+			if(F.fluid_amount <= FLUID_DELETING)
+				qdel(F)
+			else
+				F.update_position_and_alpha()
 		scheck()
 	for(var/thing in active_fluids)
 		var/obj/effect/fluid/F = thing
-		F.update_overlays()
+		if(F) F.update_overlays()
 		scheck()
 
 	return 1
